Add files via upload
Added new font for numbers because MonsterTitle only had the numbers 0,1,2. Distance should now show 0 at the beginning. Created a countdown for the beginning of the game. Updated the pause menu to make it interactive and able to control with mouse as well as arrow keys. Updated Game Over screen.